Fixed Bug #{3646}
Done.

The detail browser now has a page for the econ model and a corresponding icon in the Objects tree that looks like a dollar sign.

That page simply reports on the condition of the econ model and if there is a problem, some detail about the problem (this area should become more robust moving forward).

Now, if the econ model encounters an error in trying to find a solution to the CGE a dialog box pops up informing the user that this is what has happened, the econ model is *automatically* disabled, and the user given the option of going directly to the Econ detail browser page.

There should no longer be a bgerror thrown from the econ model after Athena has come up.

Suggested tests:
 * Inspect the detail browser and, in particular, the content of the Econ page.
 * Disable the econ model by setting econ.disable to true and verifty that the Econ page reports this.
 * Set all the revenue in the "actors" row in the SAM to 0.0 and lock the scenario (this is an error condition).
 * See that the message box described pops up and that the detail browser provides some detail on the problems.