Fixed Bug #{3669}
The lions share of changes is in support of actor's income becoming an input
to the CGE. New cells needed to be defined on the In page for this.

Another new input is graft, since this is computed from all actor's share of
graft from foreign aid to the region.

Consumer security factor is redefined to be the fraction of consumers UN-afraid
to go out and purchase goods. It will need to be computed this way in the 
econ model.

Other changes are more cosmetic, like global changes to cell names to be in line
with Bob's theory paper.

Testing of this will be in concert with testing of the actor income/revenue 
interface to Athena.