Fixed Bug #{3679}
Done!

The defaults for the affected parameters (econ.shares.*) are all changed so that they are now number of shares.
The fractions computed from the default shares are the old defaults when these parameters were fractions.

To test:
Change the econ.shares.* parameters for various expenditure classes and verify that money is allocated correctly for each expenditure class whether it be a tactic or overhead.