Fixed Bug #{3686}
Done!

For the first problem Athena and the CGE needed some bullet proofing to deal with the fact that there may be no actor income or expenditures.

For the second problem a flag is introduced in the 6x6 cell model that Athena can set to inform the CGE that the actors sector does not get black market profits.

To test:
* Create a scenario where actors have no income/cash reserves/cash on hand and to tactics that spend money. Also set all the econ.share.overhead.* parameters to zero.
* Lock the scenario and verify that the actors sector in the CGE is completely zeroed out and that Athena will still advance.

These are sufficient to test both changes above. Feel free to try different combinations.