Fixed Bug #{3691}
athena_cell(1) uses a state machine to know whether the model has
been edited and whether it has been validated since it was edited.  It is
necessary to execute the state machine just prior to checking or solving
the model; this clears the "modified" flag.  Otherwise, the model will
still appear to have been modified even after the check or solution is
done.