Fixed Bug #{3695}
Done!

Changing remittances in the SAM now causes the money flowing to the pop sector to increase, which increases their demand from other sectors.

I also changed the variable that contains this value in the SAM from RFN to REM to more accurately reflect what it represents, namely, remittances only.