Fixed Bug #{3698}
Fixed!

Bring Athena up and change some values in the SAM. Then before and after locking the scenario, select "New Scenario..." from the File menu.

Verify that when the new scenario is initialized the SAM data goes back to the defaults, which are whatever was in the cell model originally.