Fixed Bug #{3702}
Fixed!

There is an on-lock sanity check that warns the user if net revenue in the black sector is negative.

If the user decides to carry on anyway then the econ model does not have any black market revenues go to actors, if they get shares of black market net revenue or to the world, if no actors get black market profits.

This prevents what amount to the actors (or the world) subsidizing the black market.


To test:
Set the price of black market feedstock to a high number (in my testing $2M was enough) and lock the scenario.
Verify that a sanity check failure occurs with an appropriate message.
Click continue and verify that the CGE solves, but no actors get any income from black market profits (you can use the detail browser to verify this).
If no actors have any income from black market profits (ie. non have shares of it) then verify that the world does not get stuck with the bill.