Fixed Bug #{3706}
Done!

The new order is ECON:CGE:UPDATE and can be used to change cell values in the CGE.
An order dialog for this order can be brought up from the Orders > Economics menu. 
The order can only be sent when in the PAUSED or TACTIC state.

To test:
Bring Athena up and lock it. Use the order to change a cell value, the one that Rob is particularly interested in is called PF.world.black and it affects the net revenues in the black market.
Change the value of that and verify that the net revenues change.

The test suite and docs are updated as well.