Fixed Bug #{3714}
Done!

This was requested by Rob and is just a minor enhancement to the model. 

The SAM tab now has an input for subsistence wage which is a pass through to the CGE. The CGE uses that wage and the number of people engaged in subsistence agriculture from the demographics model to compute the contribution to GDP.

To test:
Vary the subsistence wage on the SAM and lock the scenario. Verify that the GDP and per capita GDP, which is also new, changes with the wage. These can be found under the outputs section on the Econ > Overview tab.

The helps are also updated.