Fixed Bug #{3717}
Realized belatedly that ComputeForceActivityFlags was computing the security
flag for all n,g,a combinations, whether any troops were assigned to the 
activity or not.  If we assume that the security flag is 0 for n,g,a
if no troops are allocated (in which case it doesn't matter) then we only
need to compute it for cases that matter, resulting in a major speed up.