Fixed Bug #{3719}
Defined stance_nfg_only_view, which only shows the stance if it's 
been explicitly set in some way; unlike stance_nfg_view, it doesn't pull in 
uram_hrel.  That turns out not to be useful when computing security, because
if there's a stance you need to compute an effective relationship, and if 
there's not you've already got the hrel.