Fixed Bug #{3724}
Found another problem and fixed it.

The logic for choosing which query to use to generate the sigevents table had a bug. Basically, any query with a single tag (ie. a neighborhood ID) used the logic for multiple tags.

The data returned up to now has always been correct even though the queries differ slightly. But the query that should have been used returns the data faster.

The test in Comment 1 applies to this fix as well.