Fixed Bug #{3741}
Done!

A.goods.pop is now computed each time the CGE is solved. This will account for a change in remittances and/or a change in the number of consumers.

The Overview tab under Econ is also updated to include the per capita demand for goods as an output in the "Other Outputs" section.

The helps are also updated.

To test:
Bring up Athena with the econ model enabled and lock the scenario. Go back to prep and change the remittances on the SAM tab.
Verify that when you lock the scenario the GDP (and maybe the unemployment rate) change along with the per capita demand for goods.