Fixed Bug #{3747}
Fixed bug in cif.tcl when re-entering the time stream after loading a
snapshot.  cif.tcl assumes that the "undo" column is "" if the order
cannot be undone; but after a snapshot is loaded, it will be NULL in 
this case.  This fix makes a simple change to the test that accepts 
either "" or NULL; a better solution would be to clear the column to NULL
always.