Fixed Bug #{3760}
Done.

Actors income is now multiplied by 52 before setting the revenue row in the SAM and, thus, the CGE. Actors expenditures as a result of tactics (or overhead) are now multiplied by 52 before setting them in the actors expenditures column of the SAM (at lock) and CGE (during run).

While running actors revenue in the CGE is divided by 52 before giving it to the actors.

Suggested tests:
Allocate an actor some income from one or more sectors.
Upon lock, the SAM should show the amounts allocated in a years worth of money, it should also show a years worth of expenditures.
The CGE should also show a years worth of money (albeit scaled to the actual population).
When actors receive income it should be based on 1/52 of the amount allocated to them from the CGE.