Fixed Bug #{3761}
Done!

The default SAM needed to be updated with new values for remittances and base consumers.

The goods and black sector shortages are now in line with long term demand vs. short term supply.

The pop sector is a little more complicated, however, since there is the issue of labor security.

The long term demand in the pop sector is what is desired by all other sectors and the short term supply is what is available given turbulence and security.

However, shortage in the pop sector is based on capacity (ie how many people could possibly work) vs. short term supply. So the shortage is not necessarily long term demand vs. short term supply.

Suggested tests:
 * To increase long term demand in the goods sector (if it doesn't already top the short term supply), drop the base price and you should see the quantity demanded in the "Quantities" matrix on the  Overview tab display the long term demand which should be the sum of the value in the "Quantity" column for that sector and the shortage in the goods sector.
 * You can do the same for the black sector.
 * The pop sector can be affected the same way, but the shortage is not simply the difference for the reasons laid out above.