Fixed Bug #{3734}
This should now be fixed.

In cases where a particular sanity check can be done it is performed. For instance, the goods and pop sectors must have revenues and expenditures (otherwise, what's the point?)

For cases where it can't, then Athena should just gracefully inform the user that the econ model is having problems and must be disabled.

Suggested tests:
* Zero out the goods sector and try to lock, you should see a sanity check failure.
* Do the same for the pop sector.
* Try zeroing out the entire SAM.
* Try zeroing out other sectors.