Fixed Bug #{3758}
Deleted the ability to save the current model parameter settings to a 
"default model parameters" file.  This feature was added early in the 
development of Athena 1.0, when it wasn't clear that model parameter values
would be saved as part of the scenario file.  

It no longer worked because it explicitly referenced a Linux-specific file 
path, thus causing the bgerror; rather than fix that, we chose to remove
the feature altogether, because its only real purpose at this point is
to cause unexpected surprises.

In particular, I:

* Removed the Save and Clear Defaults items from the File/Parameters menu.
* Removed the "parm defaults" family of executive commands.
* Removed the code that reads defaults from the file (if it exists) from
  scenario.tcl.
* Removed the related code from the parm(sim) and parmdb(n) modules.
* Updated the on-line help.