Fixed Bug #{3737}
appwin(sim) was assuming that the CLI widget was managed by the panedwindow
widget if the CLI was mapped, and not otherwise.  However, when the sash
is pulled all of the way down so that the CLI widget is not seen, it is 
simultaneously managed by the panedwindow but unmapped; hence the error.

However, the panedwindow widget has a subcommand, "panes", that lists the 
names of the managed widgets.  If the CLI widget is listed, then it is
managed; otherwise not.  This fix updates appwin(sim) to do this look up,
instead of seeing whether the CLI is mapped or not.