Fixed Bug #{3363}
The [messagebox popup] dialog provides an -onclose option, that specifies 
which of the buttons to mimic when the dialog's "x" button is pressed.  It
wasn't being used; and the default behavior is to trigger the left-most
button.  As this is usually the OK button, comedy ensues.  This package
updates the [messagebox popup] calls in athena_sim to use -onclose.