Fixed Bug #{3450}
The problem was that gui_nbhoods assumed that if control_n.controller is 
NULL, then it should use nbhoods.controller, and used coalesce() to do this.

We use this pattern when we have two tables, one with scenario data and one
with simulation data, to pick up the simulation data when locked but use the
scenario data when unlocked.  The pattern is great so long as a NULL in the
simulation data table means that there's no simulation data.  In this case,
however, NULL is a valid simulation value; we can't use NULL in the control_n
table to imply that the scenario is unlocked.

Therefore:

* Added "locked()" SQL function for use in GUI views.
* gui_nbhoods now defines the controller column using nbhoods.controller
  when unlocked and control_n.controller when locked.