Fixed Bug #{3806}
Done.

The GDP is displayed on the Outputs section of the SAM tab. Also included is the contribution of subsistence agriculture to the GDP by means of the base number of subsisters and their base wage.

To test:
Bring up Athena and inspect the SAM tab for the fields described. Verify that as you change the pertinent fields in the SAM matrix the GDP changes accordingly (see Sec. 4.6.4 of the Econ paper for the pertinent fields)