Fixed Bug #{3819}
Fixed.  Add additional test:

Create a scenario with at least two neighborhoods.  Deploy a force group to 
the first but not the second, and lock the scenario.  Verify by looking at
the force_ng table that the group has zero personnel and a zero security in
the second neighborhood.

Then write tactics such that at time 10 some troops are deployed to the
second neighborhood, and that at time 11 they are not.  Verify that the
force group's security in the second neighborhood is non-zero at time 10 
and zero at time 11.