Fixed Bug #{3820}
This modification includes an update to a lot of the internal narrative, it's gotten way out of date.

There is also a basic change to the structure: the medium-run (M) and short-run (S) pages are now copied from the Calibration (Cal) page rather than M being copied from L and S from M. L continues to be copied from Cal as it was before.

This makes the computation of the constrained economy more straight forward.

The cell model still functions like it did before, so there should be no noticeable difference in output.

More to come.