Fixed Bug #{3823}
Done.

This was a geniune bug and an obvious oversight. Yuk!

Now, when the scenario is locked the strategy module executes on-lock tactics and, hence, accounts for actors baseline spending before the econ module initializes.

The econ module now picks up the correct figures for the actors' sector base case.

To test:
Create a scenario that has actors executing on-lock strategies that expend money. When locking, verify that the combination of money spent on overhead and on on-lock strategies is accounted for in the actors sector column.

The actors sector revenues should be the same as before.