Fixed Bug #{3825}
Done!

Create at least one empty group. Create a FUNDENI tactic that tried to fund that group(s). 
There should be no error when the tactic fires, and no cash should be spent.
The sigevents log should also show that the funding could not occur because the group(s) are empty.

The automated test suite has been updated to test this case and the Test Executive Deputy has been updated to have a CIV group that is empty.