Fixed Bug #{3827}
Done!

When the scenario is locked a set of radio buttons will appear on the Econ/SAM tab that allow the user to switch between the three views of the data: input flows, balancing flows and base SAM.

The input flows include the data input directly into the matrix and into actor and strategy definitions.
The balancing flows are computed from input flows and include the values found in section 6.1 of the econ paper.
The base SAM is the sum of the input and balancing flows.

The helps have also been updated.

Suggested tests:
* After scenario lock, inspect the GUI and try the three different views.
* Verify that each view is as described.