Fixed Bug #{3830}
Updated all of the various kinds of driver (except for MADs, each of
which already gets a distinct driver ID) so that the driver IDs are
assigned by means of signatures.  Over time we'll be able to simplify the
code further; the driver ID can be looked up in the rule set itself, rather
than being carried in some other entity.