Fixed Bug #{3837}
Done!

The Econ/SAM tab now has a spot to specify an annual percentage change in remittances.

This value is used by Athena to compute a new value for remittances each time tick. This new value is then updated in the CGE and a new equilibrium computed.

This annual percentage can be changed on the fly through an order available under the Orders > Economics menu or through the use of an EXECUTIVE tactic.

The annual percentage can be positive, negative or zero (default) and is converted to a weekly fractional increase.(Ex: 6% = 6/100/52 change per week.)

Suggested tests:
* Bring up Athena and specify a positive change rate. Verify that when locked, the CGE tab shows this value and that when time advances, remittances increase resulting in a new equilibrium.
* Try a negative change rate and verify a decrease in remittances.
* Use the Orders > Economics > Update Remittance Change Rate... menu to change the remittance rate.
* Verify that the change rate appears on the CGE tab (it won't on the SAM tab, and it shouldn't) and that the increase/decrease in remittances changes accordingly.