Fixed Bug #{3841}
Done!

The SAM also has defaults that are in line with the latest econ study.

The Athena econ module is updated to send base unemployment and the base work force capacity to initialize the CGE.

Suggested tests:
Bring up Athena and inspect the SAM tab. You should see a base GDP around $221B if the defaults are correct.
Lock the scenario and verify that there are no problems with CGE initialization.

Keep in mind that at the time of this bug commit the CGE and SAM do not yet match, so that may not be the case when this is tested.