Fixed Bug #{3850}
Done.

To test:
Create a scenario that has around 160M people and 90M laborers. 

Lock the scenario and see if there are errors. 

Try different sized scenarios. The CGE should be robust enough to scale the economy.

Still to be done:
Have the CGE output match the SAM. 
The first step was to have the Cal page match the SAM, and that's working but there's no insight to it, yet...