Fixed Bug #{3859}
Added the new lfp and housing columns to the civgroups table, and a new
enumeration, ehousing(n), to projtypes(n).  Updated the civgroup(sim)
mutators and orders, and the test suite.  Added a new "fillparms" routine
to helpers(sim) that does for other entities what [tactic delta] does for
tactics.