Fixed Bug #{3856}
The scaling problem is now fixed.

There is a caveat, however, when the scenario has an extremely small population (maybe less than 10k total) and the default numbers in the SAM are not changed.
There are two new econ parameters called econ.gdpExp and econ.empExp which correspond to exponents in the CGE cell model that affect how the calibration of the model converges.
When the population is extremely small, the Cal page will fail to converge which requires setting these parameters to zero.

Since the default SAM is for an economy that has many millions of people, it stands to reason that trying to use those numbers with such a small population is nonsense.

Nonetheless, sometimes testing with a small scenario is necessary, so setting these to zero will allow testing with the default SAM.

There are also views of the economy available on the Overview tab: Capacity Constrained with Security Effects, Capacity Constrainted, Unconstrained and Calibrated. They are as follows:

Capacity Constrained with Security Effects: The economy as constrained by the available workforce, the production capacity in the goods sector and both consumer and labor security factors.
Capacity Constrained: The economy constrained only by labor and goods production capacity.
Unconstrained: The economy with all available workers working and goods production maximum.
Calibrated: This should match the Capacity Constrained page, but is really for debugging at this point, this selection may go away in the future.

The test suite has also been updated.

Suggested tests:
Try creating scenarios that have less (half or one quarter the size) of the default SAM size. The default SAM is on the order of 200 million with 40 million of those being subsistence farmers.
Lock the scenario and see that the GDP of the "Capacity Constrained" economy on the Overview tab matches that on the SAM tab.

Try other sized scenarios.
You can also try tiny scenarios with the parameters above set to zero, but the numbers in the econ model will be nonsense, there just shouldn't be any bgerrors.

Insepect the helps for the Econ tab.