Fixed Bug #{3884}
Fixed.

Rob Crowson also indicated that in some scenarios, remittances should be treated as taxable income while in others it should not.

A new parmdb parameter is introduced that controls whether remittances are treated as before tax income or after tax income. The parameter is called econ.REMisTaxed and has a default of "yes".

Tests:
Try different unemployment rates in the SAM and when the scenario is locked, the unemployment rate in the "Calibration" and "Capacity Constrained" views should match.
Use the new parameter to have remittances be treated as before and after income tax.
For sectors 'i' that have a non-zero t.i.pop int shape parameter matrix on the SAM tab, the money flow in the corresponding CGE cell on the overview tab should be higher if REM is taxed.