Fixed Bug #{3777}
Done.

The per actor, sector-by-sector expenditures are now visible in two places: the detail browser and a new table browser under the Econ tab labeled "Actors"

In the detail browser, expenditures are shown in a single table for each actor. The table has two rows, one for expenditures made during the current time step and one for expenditures made to date.
Each row is broken down by economic sector.

In the Econ > Actors tab there are three views of expenditures available using the "Views:" dropdown box on the tab. The "This Week" view shows the expenditures made during the current time step, the "To Date" view shows total expenditures made since t0 and "Annualized" shows the current expenditures multiplied by 52.0
The total of the annualized expenditures for all actors should match the column of the actors sector in the Econ > Overview tab.

The automated test suite has been updated as well.

Suggested tests:
* Create a scenario with both Budget actors and Income actors.
* Give them strategies that expend money, including the "Spend" tactic.
* Verify that the detail browser and Econ > Actors tab are consistent with the expenditures specified in the tactics, given the actor has enough money to perform them.
* Inspect the help for the Econ > Actors tab.