Fixed Bug #{3888}
Done.

The parameter is called econ.turFrac and can have a value between 0.0 and 1.0 (although a turbulence of 1.0 is ridiculous).

If the unemployment rate in the SAM is not greater than econ.turFrac a sanity check failure will occur.

Suggested tests:
* Nominal modes: Unemployment Rates of 8% to 15% with turbulenc fractions in the 0.02 to 0.08 range.
* Error modes: Unemployment Rates less than turbulence (eg. 4.0% and 0.08)
* Insane modes: Eg., unemployment rate of 100%, turbulence of 1.0. (This is very ill-conditioned and should fail convergence gracefully).

To see that econ.turFrac is having an effect, look at the Idle Capacity in the pop sector on the Econ > Overview tab. Those in turbulence are not part of the labor capacity, so the lower the turbulence, the higher the idle capacity given that unemployment > turbulence.