Fixed Bug #{3890}
Done.

The "Other Inputs" section on the Econ > SAM tab has this.

As production capacity in the black market approaches than the quantity 
demanded, you should see a reduction in idle capacity on the Econ > Overview 
tab for the capacity constrained views.

Likewise, if production capacity is below the quantity demanded you should see
an increase in shortages there.