Fixed Bug #{3891}
Done!

The cash module now allocates income to actors on lock. The cash on hand is not,
however, a constraint when it comes to on-lock tactics. All on-lock tactics still
happen regardless of cash on hand. Only when the SPEND tactic is executed with
"all remaining cash" specified does cash on hand come into play. Whatever is
left in cash on hand is allocated to the sectors according to the profile 
provided in the tactic.

To test:
* Create an actor with a strategy that has tactics that spend money. 
* Verify that when the SPEND tactic is executed with "all remaining cash" specified, the correct amount of money is shown in the expenditures row of the SAM matrix.
* Verify that if all money is spent on other tactics, a SPEND tactic like this does not increase the amount of money in the expenditures row.
* Verify that if the SPEND tactic has a dollar amount, then that money is allocated to sectors on scenario lock regardless of cash on hand.