Fixed Bug #{3858}
Done.

This was a minor effect, but an important one. When the revenue in sectors that
are taxed by actors change, then each actors income changes. Which, in turn, 
changes the revenue by the entire sector.

The CGE does this in standalone, but not setting them from Athena when the CGE
is running in that context is wrong.

It's sufficient to see that the econ model runs fine in other tests that use
the actors sector.

This effect is seen greatest when an actor uses the SPEND tactic with "all
remaining money" as the choice. The economy will actually expand (assuming an
actor has money left over) because actors spend more on other sectors, which 
then have greater revenues, which then are taxed by actors who get more and 
then spend more. But the expansion is not infinite, there is a bound at which
expansion will stop.

Who says tax and spend doesn't work :-^