Fixed Bug #{3894}
Done.

The need for this came about after tracking down a problem for TRISA. Basically,
an actor was defined as having *way* to much income as compared to the rest
of the economy. Because of this, some shape parameters were just meaningless,
but Athena doesn't know this until it attempts to compute the actors sector and
solve the CGE, which fails.

When any CGE solution failure occurs Athena does this:
* Pops up a dialog telling the user there has been a problem
* Disables the econ model
* Has the econ model produce a report
* Finishes locking

The report that gets generated in step 3 has been beefed up to look at some 
possible problems and then offer troubleshooting suggestions. Right now there
is only one suggestion implemented, there may be more moving forward.

To test:
Create a scenario that has an actor that has a ridiculous amount of income from
one or more sectors as compared to the size of the economy, like 100 times more
than the GDP.

Lock the scenario. The econ model should fail because of the CGE and perform
the steps above. Inspect the report for an appropriate trouble shooting
suggestion.