Fixed Bug #{3898}
No, it wasn't.  The rule set was implemented, and the Athena test suite
tested it by calling it directly; but the application itself actually 
never invoked it.  (Sigh!)  This is now fixed, as are the test suite changes
required because the CONSUMP rule set now fires when it didn't used to.