Fixed Bug #{3892}
The code for the CIVGROUP:UPDATE:MULTI order got very complicated when we
added the "housing" attribute, due to the interactions between the sa_flag
and "housing" attributes.  At the time, we considered removing the capability
altogether, because it isn't often useful; there are few attributes
you'd want to set the same for a set of distinct civgroups.

Consequently, following the philosophy that you should delete bad code 
rather than fixing it, we're removing the capability.  It is no longer
possible to edit multiple civgroups at the same time.