Fixed Bug #{3905}
The 'autogen' module is introduced along with the 'autogen' family of commands for automatic scenario generation.

The help documentation for Executive Commands has been updated.

To test:
View the 'autogen' helps in the detail browser under Executive Commands
Exercise all of the commands that are available and verify that they work as described.
Try error cases such as creating civilians before there are any neighborhoods.