Fixed Bug #{3911}
The user can now explicitly set and browse the "natural" cooperation level
for each cooperation curve.  In addition, SIM:REBASE sets only the
initial baseline, so that that natural cooperation level is unchanged
after a rebase.  Updated the test suite and the help pages for 
the COOP:* orders and the Attitudes/Cooperation tab.