Fixed Bug #{3933}
Done.

The rebasing of the econ model should work correctly now.

There is a new order introduced under the Order > Economics menu that allows the user to adjust historical data from a rebase, if needed.

Tests:
Create a scenario and run for a while. Take note of the values in the CGE, particularly unemployment and the GDP. Rebase the simulation and immediately lock.
Verify that the CGE is the same from the end of the previous run.

Rerun the test above but with a non-zero remittance change rate. Do the same with at least one CIV group with a non-zero population growth rate.
Try a test with both of those non-zero.

Each time the rebased scenario, when locked, should look like the end state of the previous run.

Use the Order > Economics menu to change some of the key historical parameters.
Verify that changing of parameters can be undone.

Inspect the helps for the new order.