Fixed Bug #{3944}
Snapshots are identified by the sim time tick at which they were saved.
The [sim snapshot first] call, which is triggered by the "|<<" button on 
the toolbar, was explicitly requesting the snapshot for t=0, rather than
the first snapshot in the list of snapshots, which is now always taken
at t=start_tick.  Fixed.