Fixed Bug #{3965}
Fixed!

The civilian ID is now just an integer that counts upward monitonically preceded by a 'C'.
The max number of civ groups is 10000, but you'd be crazy to create that many.

To test:
autogen a scenario with 100 nbhoods and 10 civ groups (this may take a while). Verify that all civ groups are successfully created.