Fixed Bug #{3960}
Done.

You should be able to use the "send" command with the -g parameter specified, but it has absolutely no effect whatsoever.
This is being put in for the benefit of TRISA, so their order scripts still work during the current round of studies.

Tests:
Create ensits both by the command line and via the GUI. Verify that you can pass in the -g parameter with any string you wish.
Update ensits by both methods and verify that -g works in that case, too.