Fixed Bug #{3916}
Done.

SAT:UPDATE and SAT:UPDATE:MULTI order dialogs now have the combobox for
saliency.

To test:
Bring up Athena with some civ groups and bring up the update satisfaction order dialog.
Verify that next to the saliency slider is a combobox that shows the text description of the level of saliency.
Use the slider to change the combobox value.
Use the combobox to change the slider value.

Do these tests for both single and multi-selected civ groups.