Fixed Bug #{3981}
Done.

Attached to this bug is an order script that can be loaded into an empty scenario to create the conditions to test this bug.
It sets up the actors sector in such a way that failure is assured.

Read in the order script using the order command and lock the scenario. Prior to this fix, a bgerror would be generated. Now you should see in the detail browser more information about what the problem is.