Fixed Bug #{3993}
This commit introduces the CURSE object and associated inject objects used to cause attitude changes when a CURSE occurs.

CURSEs occur as the result of the user adding a yet to be implemented CURSE tactic to the SYSTEM agents strategy.

The concept of inject "roles" is also introduced. Roles can be thought of as variables that contain any number of groups or actors depending on the type of inject: HREL, VREL, SAT or COOP.
When the CURSE tactic is specified, roles are assigned appropriate lists of entities based on the inject type. When the tactic executes, the effects of injects are applied to the groups or actors that comprise those roles.

The helps have also been updated with the new objects.

Suggested tests:
* Read the documentation for the new simulation objects: CURSEs and injects. In there see how roles come into play.
* Create some CURSEs and associate some injects with them using the new CURSE browser under the Attitudes tab.

Note that CURSEs do not yet execute since the tactic that uses them has not yet been implemented.