Fixed Bug #{3999}
This commit has the documentation addition to the detail browser. 

In the "Objects" tab there is now a CURSEs object on the tree that gets populated when CURSEs are created.
CURSE objects have their injects associated with them and, if a CURSE tactic executes, the associated inputs to DAM are shown.