Fixed Bug #{4014}
The validators are in place and the automated tests that were skipped are now being executed.

A couple more tests were added to the test suite.

To test:
Use the 'send' command for TACTIC:CURSE:CREATE and TACTIC:CURSE:UPDATE and try weird combinations of things for the "roles" field.

Verify that proper error messages are generated.