Fixed Bug #{4028}
Fixed!

As part of this commit the narrative for CURSE tactics have been shortened to show only the roles and their values. If a user wants to see the specifics of a particular curse (magnitude, etc...) he or she will need to see the CURSE browser.

Retest and inspect CURSE tactic narratives.