Fixed Bug #{3946}
Most everything described is fixed.

As far as undoing individual autogens that will probably not be supported. 
Autogen is designed for use in enabling testing, it would never be used to create a "real life" scenario so supporting undo of autogen commands just doesn't make sense.

If you want to completely undo an autogen just type "new" at the command line.

Retest and inspect the "autgen attroe" help.