Fixed Bug #{4042}
Done.

athena_test(1) has an option: "-notk" that will run the test suite without Tk and, thus, a GUI being loaded.
As a result, some tests in the 020-MAP test module have constraints to only run when Tk is loaded since those tests need it.

The CM plan is updated with an optional step to run the test suite with this option.