Fixed Bug #{4045}
Done!

Create some CURSEs with injects and use the CURSE tactic to specify which groups are affected.
You should see that gofers are now used to specify the rules under which a role is filled when the tactic executes.

The helps for the CURSE tactic roles has also been updated to link to the gofer documentation.

Suggested tests:
* Use different gofers to fill in roles for CURSE tactics
* Test that if all of the injects for a particular curse have roles that remain unfilled because the rule to fill it is not met, then the tactic will not execute.
* Test that if at least one inject is executed, then the tactic is shown as having been executed.

For injects that do not execute because a gofer returns an empty list, a "detail" level log message is generated.