Fixed Bug #{4052}
Fixed.

The easiest way to test this is to bring up a working scenario in the GUI and 
then make it invalid. There are a number of ways to do this one of which is to 
delete a group that is referenced by some tactic (CURSE or FUNDENI comes to 
mind) and then try to lock causing an on-lock sanity check failure.

Save the invalid scenario and then use it in batch mode and have Athena advance.

Verify that an error.log is created with a description of the problem encountered.
On Linux machines you will also get output at the terminal window directing
you to look at error.log.

This commit also includes a check in the econ model for failures in solving the
SAM or the CGE and if there's no GUI present, writing the problem to error.log.

To get the SAM to fail on lock set the Base Unemployment Rate to 0.0
To get the CGE to fail on lock set Subsistence Wage in the SAM to a ridiculously
large number (I used $100T).