Fixed Bug #{4041}
This has been fixed.

If an attempt is made to create an inject with "Define new role" 
selected for new injects or "Rename role" selected for existing 
injects AND the role name exists in any previously created inject,
the validation of the role fails ensuring the user must specify
a brand new role name.

The behavior for using existing role names is unchanged.

The test suite has been updated and the helps for the INJECT:* family of
orders have been updated with the parms used to specify whether roles are
new or existing.