Fixed Bug #{4071}
Fixed.

When a failure to solve the CGE occurs (which is more likely) in batch mode the
problem is reported to the on-tick sanity checker which allows Athena to stop
gracefully when running in a cluster environment.

If a failure to solve should occur in the SAM in batch mode it would be as part
of an on-lock sanity check. This also allows Athena to stop gracefully.

Note, however, that this is less likely problem in a cluster environment since 
presumably the user has already created a base scenario without problems.