Fixed Bug #{4072}
This is now fixed.

Only roles already defined within the same CURSE are available as exiting roles when adding new injects.

There is still another problem, however, but this is captured in Bug 4073.

To test:

Verify that when adding injects to a CURSE, only roles already defined for that CURSE are available as existing roles.